% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{logistic_orbit}
\alias{logistic_orbit}
\title{Logistic map orbit}
\usage{
logistic_orbit(r, x0 = 0.2, n = 100L)
}
\arguments{
\item{r}{Growth parameter in \verb{[0, 4]}.}

\item{x0}{Initial value in \verb{(0, 1)}.}

\item{n}{Number of iterations.}
}
\value{
Numeric vector \verb{x_0, ..., x_n}.
}
\description{
Iterates \eqn{x_{n+1} = r x_n (1 - x_n)}.
}
\examples{
plot(logistic_orbit(3.9, n = 100), type = "b", pch = 16, cex = 0.6)
}
