% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot.chaos_map}
\alias{plot.chaos_map}
\alias{plot.bifurcation}
\title{Plot a map attractor or a bifurcation diagram}
\usage{
\method{plot}{chaos_map}(x, col = "#1f2a44", pch = ".", cex = 1, ...)

\method{plot}{bifurcation}(x, col = "#1f2a44", pch = ".", cex = 1, ...)
}
\arguments{
\item{x}{A \code{chaos_map} or \code{bifurcation} object.}

\item{col}{Point colour.}

\item{pch, cex}{Point symbol and size; the default is a single pixel.}

\item{...}{Passed on to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly, \code{x}.
}
\description{
Base-graphics methods that draw hundreds of thousands of points quickly.
For publication-style output see \code{\link[=plot_bifurcation]{plot_bifurcation()}} and
\code{\link[=plot_standard_map]{plot_standard_map()}} (ggplot2).
}
