% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attractor}
\alias{plot_attractor}
\alias{plot.chaos_trajectory}
\title{Plot a trajectory as a 2D phase portrait}
\usage{
plot_attractor(x, vars = c("x", "z"), col = NULL, lwd = 0.5, ...)

\method{plot}{chaos_trajectory}(x, vars = c("x", "z"), col = NULL, lwd = 0.5, ...)
}
\arguments{
\item{x}{A \code{chaos_trajectory} (from \code{\link[=lorenz]{lorenz()}} or \code{\link[=roessler]{roessler()}}).}

\item{vars}{Two variable names to draw, e.g. \code{c("x", "z")}.}

\item{col}{Line colour, or a vector of colours (one per row) to encode time.}

\item{lwd}{Line width.}

\item{...}{Passed on to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly, \code{x}.
}
\description{
Plot a trajectory as a 2D phase portrait
}
\examples{
plot(lorenz(n = 8000))
}
