% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_cobweb}
\alias{plot_cobweb}
\title{Plot a cobweb diagram}
\usage{
plot_cobweb(cw, ...)
}
\arguments{
\item{cw}{A \code{cobweb} object from \code{\link[=logistic_cobweb]{logistic_cobweb()}}.}

\item{...}{Unused.}
}
\value{
A ggplot.
}
\description{
Plot a cobweb diagram
}
