% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_standard_map}
\alias{plot_standard_map}
\title{Plot the standard-map phase portrait with ggplot2}
\usage{
plot_standard_map(sm, size = 0.15)
}
\arguments{
\item{sm}{Output of \code{\link[=standard_map]{standard_map()}}.}

\item{size}{Point size.}
}
\value{
A ggplot.
}
\description{
Plot the standard-map phase portrait with ggplot2
}
